@@ -29,8 +29,8 @@
     "@tanstack/react-query": "^5.18.0",
     "dotenv": "^16.3.1",
     "globals": "^15.11.0",
-    "i18next": "^23.7.3",
-    "monaco-editor": "^0.51.0",
+    "i18next": "^24.0.2",
+    "monaco-editor": "^0.52.0",
     "react": "^18.2.0",
     "react-colorful": "^5.6.1",
     "react-dom": "^18.2.0",
@@ -50,7 +50,7 @@
     "@thedutchcoder/postcss-rem-to-px": "^0.0.2",
     "@total-typescript/ts-reset": "^0.6.0",
     "@types/archiver": "^6.0.1",
-    "@types/chrome": "^0.0.273",
+    "@types/chrome": "^0.0.280",
     "@types/eslint__js": "^8.42.3",
     "@types/node": "^22.1.0",
     "@types/react": "^18.2.37",
@@ -85,7 +85,7 @@
     "tailwindcss": "^3.3.6",
     "ts-json-as-const": "^1.0.7",
     "ts-node": "^10.9.1",
-    "typescript": "5.5",
+    "typescript": "5.7",
     "typescript-eslint": "^8.8.1",
     "vite": "^5.4.8",
     "zod": "^3.22.4",
