@@ -1,16 +1,16 @@
-import eslintPluginReact from "eslint-plugin-react";
-import eslintPluginNoSecrets from "eslint-plugin-no-secrets";
-import eslintPluginImport from "eslint-plugin-import";
-import eslintPluginTailwindCSS from "eslint-plugin-tailwindcss";
-import eslintPluginPromise from "eslint-plugin-promise";
-import eslintPluginPerfectionist from "eslint-plugin-perfectionist";
-import eslintPluginReactHooks from "eslint-plugin-react-hooks";
-import eslintPluginPrettier from "eslint-plugin-prettier/recommended";
-import eslintTypeScriptParser from "@typescript-eslint/parser";
-import eslintJavascript from "@eslint/js";
-import typescriptEslint from "typescript-eslint";
 import { fixupPluginRules } from "@eslint/compat";
+import eslintJavascript from "@eslint/js";
+import eslintTypeScriptParser from "@typescript-eslint/parser";
+import eslintPluginImport from "eslint-plugin-import";
+import eslintPluginNoSecrets from "eslint-plugin-no-secrets";
+import eslintPluginPerfectionist from "eslint-plugin-perfectionist";
+import eslintPluginPrettier from "eslint-plugin-prettier/recommended";
+import eslintPluginPromise from "eslint-plugin-promise";
+import eslintPluginReact from "eslint-plugin-react";
+import eslintPluginReactHooks from "eslint-plugin-react-hooks";
+import eslintPluginTailwindCSS from "eslint-plugin-tailwindcss";
 import globals from "globals";
+import typescriptEslint from "typescript-eslint";
 export default [
 	{
 		ignores: ["**/watch.js", "dist/**/*", "releases/**/*"]
@@ -22,16 +22,10 @@ export default [
 	eslintPluginImport.flatConfigs.typescript,
 	eslintPluginPromise.configs["flat/recommended"],
 	eslintPluginPrettier,
+	eslintPluginPerfectionist.configs["recommended-natural"],
 	{
-		plugins: {
-			react: eslintPluginReact,
-			"no-secrets": eslintPluginNoSecrets,
-			perfectionist: eslintPluginPerfectionist,
-			"react-hooks": fixupPluginRules(eslintPluginReactHooks)
-		},
 		languageOptions: {
 			ecmaVersion: "latest",
-			sourceType: "module",
 			globals: {
 				...globals.browser,
 				...globals.node,
@@ -44,43 +38,54 @@ export default [
 				},
 				project: "./tsconfig.json",
 				tsconfigRootDir: "."
-			}
-		},
-		settings: {
-			tailwindcss: {
-				callees: ["cn"],
-				config: "./tailwind.config.ts"
 			},
-			react: {
-				version: "detect"
-			}
+			sourceType: "module"
+		},
+		plugins: {
+			"no-secrets": eslintPluginNoSecrets,
+			react: eslintPluginReact,
+			"react-hooks": fixupPluginRules(eslintPluginReactHooks)
 		},
 		rules: {
-			"react/react-in-jsx-scope": "off",
+			"@typescript-eslint/explicit-module-boundary-types": "off",
+			"@typescript-eslint/no-explicit-any": "off",
+			"@typescript-eslint/no-floating-promises": "error",
 			"@typescript-eslint/no-unused-vars": [
 				"error",
 				{
-					varsIgnorePattern: "^_",
-					argsIgnorePattern: "^_"
+					args: "all",
+					argsIgnorePattern: "^_",
+					caughtErrors: "all",
+					caughtErrorsIgnorePattern: "^_",
+					destructuredArrayIgnorePattern: "^_",
+					ignoreRestSiblings: true,
+					varsIgnorePattern: "^_"
 				}
 			],
-			"@typescript-eslint/no-explicit-any": "off",
-			"@typescript-eslint/explicit-module-boundary-types": "off",
 			"@typescript-eslint/restrict-template-expressions": "off",
-			quotes: ["error", "double", { avoidEscape: true, allowTemplateLiterals: true }],
-			semi: ["error", "always"],
-			"prefer-const": ["error", { destructuring: "any", ignoreReadBeforeAssign: false }],
-			"prefer-destructuring": ["error", { array: true, object: true }, { enforceForRenamedProperties: true }],
-			"no-useless-escape": "off",
+			"import/first": ["error"],
+			"import/no-named-as-default-member": "off",
+			"import/no-unresolved": "off",
 			"no-empty": ["error", { allowEmptyCatch: true }],
 			"no-mixed-spaces-and-tabs": ["error", "smart-tabs"],
-			"import/first": ["error"],
 			"no-secrets/no-secrets": ["error", { tolerance: 5.0 }],
-			"import/no-unresolved": "off",
-			"tailwindcss/no-custom-classname": "off",
+			"no-useless-escape": "off",
+			"prefer-const": ["error", { destructuring: "any", ignoreReadBeforeAssign: false }],
+			"prefer-destructuring": ["error", { array: true, object: true }, { enforceForRenamedProperties: true }],
+			quotes: ["error", "double", { allowTemplateLiterals: true, avoidEscape: true }],
+			"react/react-in-jsx-scope": "off",
+			semi: ["error", "always"],
 			"tailwindcss/classnames-order": "error",
-			"@typescript-eslint/no-floating-promises": "error",
-			"import/no-named-as-default-member": "off"
+			"tailwindcss/no-custom-classname": "off"
+		},
+		settings: {
+			react: {
+				version: "detect"
+			},
+			tailwindcss: {
+				callees: ["cn"],
+				config: "./tailwind.config.ts"
+			}
 		}
 	},
 	{
