@@ -1,8 +1,8 @@
 import type { Nullable } from "@/src/types";
 import type { ChangeEvent } from "react";
 
-import { cn, debounce } from "@/src/utils/utilities";
-import React, { useCallback, useRef, useState } from "react";
+import { cn } from "@/src/utils/utilities";
+import React, { useRef, useState } from "react";
 import { IoMdEye, IoMdEyeOff } from "react-icons/io";
 
 export type TextInputProps = {
@@ -17,12 +17,21 @@ export type TextInputProps = {
 
 const TextInput: React.FC<TextInputProps> = ({ className, id, input_type, label, onChange, title, value }) => {
 	const [showPassword, setShowPassword] = useState(false);
-	const debouncedOnChange = useCallback(debounce(onChange, 300), []);
 	const inputRef = useRef<Nullable<HTMLInputElement>>(null);
+	const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
+		const { currentTarget: input } = event;
+		const { selectionStart: cursorPosition } = input;
+		onChange(event);
+		// Restore cursor position after re-render
+		requestAnimationFrame(() => {
+			if (inputRef.current && cursorPosition !== null) {
+				inputRef.current.setSelectionRange(cursorPosition, cursorPosition);
+			}
+		});
+	};
 	const handleInputWrapperClick = () => {
 		inputRef.current?.focus();
 	};
-	// FIXME: cursor not being restored to position it was in when value is saved
 	return (
 		<div aria-valuetext={value} className={cn("relative flex flex-row items-center justify-between gap-4", className)} id={id} title={title}>
 			<label htmlFor={id}>{label}</label>
@@ -44,9 +53,7 @@ const TextInput: React.FC<TextInputProps> = ({ className, id, input_type, label,
 				<input
 					className="!m-0 h-fit w-[118px] bg-transparent !p-0 !text-sm focus:outline-none"
 					id={id}
-					onChange={({ target: { value } }) => {
-						debouncedOnChange({ currentTarget: { value } });
-					}}
+					onChange={handleInputChange}
 					ref={inputRef}
 					type={showPassword && input_type === "password" ? "text" : input_type}
 					value={value}
