@@ -1,11 +1,13 @@
 import type { Nullable } from "@/src/types";
-import "./index.css";
-import { waitForSpecificMessage } from "@/src/utils/utilities";
+
 import {
 	hideOfficialArtistVideosFromHomePage,
 	observeOfficialArtistVideosFromHomePage,
 	showOfficialArtistVideosFromHomePage
 } from "@/src/features/hideOfficialArtistVideosFromHomePage/utils";
+import { waitForSpecificMessage } from "@/src/utils/utilities";
+
+import "./index.css";
 let officialArtistVideosObserver: Nullable<MutationObserver> = null;
 export async function enableHideOfficialArtistVideosFromHomePage() {
 	// Wait for the "options" message from the content script
