@@ -1,50 +1,50 @@
 export const availableLocales = [
-  "ca-ES",
-  "cs-CZ",
-  "de-DE",
-  "en-GB",
-  "en-US",
-  "es-ES",
-  "fa-IR",
-  "fr-FR",
-  "he-IL",
-  "hi-IN",
-  "it-IT",
-  "ja-JP",
-  "ko-KR",
-  "pl-PL",
-  "pt-BR",
-  "ru-RU",
-  "sv-SE",
-  "tr-TR",
-  "uk-UA",
-  "vi-VN",
-  "zh-CN",
-  "zh-TW"
+	"ca-ES",
+	"cs-CZ",
+	"de-DE",
+	"en-GB",
+	"en-US",
+	"es-ES",
+	"fa-IR",
+	"fr-FR",
+	"he-IL",
+	"hi-IN",
+	"it-IT",
+	"ja-JP",
+	"ko-KR",
+	"pl-PL",
+	"pt-BR",
+	"ru-RU",
+	"sv-SE",
+	"tr-TR",
+	"uk-UA",
+	"vi-VN",
+	"zh-CN",
+	"zh-TW"
 ] as const;
 export const localePercentages: Record<AvailableLocales, number> = {
-	"en-US": 100,
 	"ca-ES": 0,
 	"cs-CZ": 0,
 	"de-DE": 27,
 	"en-GB": 1,
+	"en-US": 100,
 	"es-ES": 48,
 	"fa-IR": 0,
 	"fr-FR": 50,
 	"he-IL": 0,
 	"hi-IN": 0,
-	"it-IT": 89,
+	"it-IT": 100,
 	"ja-JP": 89,
 	"ko-KR": 89,
 	"pl-PL": 0,
 	"pt-BR": 55,
-	"ru-RU": 89,
-	"sv-SE": 89,
-	"tr-TR": 62,
+	"ru-RU": 100,
+	"sv-SE": 100,
+	"tr-TR": 69,
 	"uk-UA": 18,
 	"vi-VN": 0,
-	"zh-CN": 89,
-	"zh-TW": 86
+	"zh-CN": 100,
+	"zh-TW": 97
 };
 export const localeDirection: Record<AvailableLocales, "ltr" | "rtl"> = {
 	"ca-ES": "ltr",
