@@ -3,15 +3,15 @@ import type { AvailableLocales } from "@/src/i18n/constants";
 import { getVideoHistory, setVideoHistory } from "@/src/features/videoHistory/utils";
 import {
 	type AllButtonNames,
+	buttonNames,
 	type ButtonPlacement,
+	type configuration,
+	type configurationKeys,
 	type ContentSendOnlyMessageMappings,
 	type ContentToBackgroundSendOnlyMessageMappings,
 	type Messages,
 	type RememberedVolumes,
-	type StorageChanges,
-	buttonNames,
-	type configuration,
-	type configurationKeys
+	type StorageChanges
 } from "@/src/types";
 import { defaultConfiguration } from "@/src/utils/constants";
 import { parseStoredValue, sendExtensionMessage, sendExtensionOnlyMessage } from "@/src/utils/utilities";
@@ -80,6 +80,23 @@ document.addEventListener("yte-message-from-youtube", () => {
 			}
 			case "request_data": {
 				switch (message.type) {
+					case "extensionURL": {
+						void sendExtensionMessage("extensionURL", "data_response", {
+							extensionURL: chrome.runtime.getURL("")
+						});
+						break;
+					}
+					case "language": {
+						const language = await new Promise<AvailableLocales>((resolve) => {
+							chrome.storage.local.get("language", (o) => {
+								resolve(o.language as AvailableLocales);
+							});
+						});
+						void sendExtensionMessage("language", "data_response", {
+							language
+						});
+						break;
+					}
 					case "options": {
 						/**
 						 * Retrieves the options from the local storage and sends them back to the youtube page.
@@ -99,6 +116,13 @@ document.addEventListener("yte-message-from-youtube", () => {
 						void sendExtensionMessage("options", "data_response", { options });
 						break;
 					}
+					case "videoHistoryAll": {
+						const videoHistory = getVideoHistory();
+						void sendExtensionMessage("videoHistoryAll", "data_response", {
+							video_history_entries: videoHistory
+						});
+						break;
+					}
 					case "videoHistoryOne": {
 						const { data } = message;
 						if (!data) return;
@@ -109,46 +133,22 @@ document.addEventListener("yte-message-from-youtube", () => {
 						});
 						break;
 					}
-					case "videoHistoryAll": {
-						const videoHistory = getVideoHistory();
-						void sendExtensionMessage("videoHistoryAll", "data_response", {
-							video_history_entries: videoHistory
-						});
-						break;
-					}
-					case "extensionURL": {
-						void sendExtensionMessage("extensionURL", "data_response", {
-							extensionURL: chrome.runtime.getURL("")
-						});
-						break;
-					}
-					case "language": {
-						const language = await new Promise<AvailableLocales>((resolve) => {
-							chrome.storage.local.get("language", (o) => {
-								resolve(o.language as AvailableLocales);
-							});
-						});
-						void sendExtensionMessage("language", "data_response", {
-							language
-						});
-						break;
-					}
 				}
 				break;
 			}
 			case "send_data": {
 				switch (message.type) {
-					case "setRememberedVolume": {
-						const { remembered_volumes: existingRememberedVolumeStringified } = await chrome.storage.local.get("remembered_volumes");
-						const existingRememberedVolumes = parseStoredValue(existingRememberedVolumeStringified as string) as RememberedVolumes;
-						void chrome.storage.local.set({ remembered_volumes: JSON.stringify({ ...existingRememberedVolumes, ...message.data }) });
-						break;
-					}
 					case "pageLoaded": {
 						chrome.storage.onChanged.addListener(storageListeners);
 						window.onunload = () => {
 							chrome.storage.onChanged.removeListener(storageListeners);
 						};
+						break;
+					}
+					case "setRememberedVolume": {
+						const { remembered_volumes: existingRememberedVolumeStringified } = await chrome.storage.local.get("remembered_volumes");
+						const existingRememberedVolumes = parseStoredValue(existingRememberedVolumeStringified as string) as RememberedVolumes;
+						void chrome.storage.local.set({ remembered_volumes: JSON.stringify({ ...existingRememberedVolumes, ...message.data }) });
 						break;
 					}
 					case "videoHistoryOne": {
@@ -271,6 +271,11 @@ const storageChangeHandler = async (changes: StorageChanges, areaName: string) =
 				automaticTheaterModeEnabled: newValue
 			});
 		},
+		enable_automatically_disable_closed_captions: (__oldValue, newValue) => {
+			sendExtensionOnlyMessage("automaticallyDisableClosedCaptionsChange", {
+				automaticallyDisableClosedCaptionsEnabled: newValue
+			});
+		},
 		enable_copy_timestamp_url_button: (__oldValue, newValue) => {
 			sendExtensionOnlyMessage("copyTimestampUrlButtonChange", {
 				copyTimestampUrlButtonEnabled: newValue
@@ -313,6 +318,9 @@ const storageChangeHandler = async (changes: StorageChanges, areaName: string) =
 				hideLiveStreamChatEnabled: newValue
 			});
 		},
+		enable_hide_official_artist_videos_from_home_page: (__oldValue, newValue) => {
+			sendExtensionOnlyMessage("hideOfficialArtistVideosFromHomePageChange", { hideOfficialArtistVideosFromHomePageEnabled: newValue });
+		},
 		enable_hide_paid_promotion_banner: (__oldValue, newValue) => {
 			sendExtensionOnlyMessage("hidePaidPromotionBannerChange", {
 				hidePaidPromotionBannerEnabled: newValue
@@ -342,9 +350,6 @@ const storageChangeHandler = async (changes: StorageChanges, areaName: string) =
 			sendExtensionOnlyMessage("maximizeButtonChange", {
 				maximizePlayerButtonEnabled: newValue
 			});
-		},
-		enable_hide_official_artist_videos_from_home_page: (__oldValue, newValue) => {
-			sendExtensionOnlyMessage("hideOfficialArtistVideosFromHomePageChange", { hideOfficialArtistVideosFromHomePageEnabled: newValue });
 		},
 		enable_open_transcript_button: (__oldValue, newValue) => {
 			sendExtensionOnlyMessage("openTranscriptButtonChange", {
@@ -395,11 +400,6 @@ const storageChangeHandler = async (changes: StorageChanges, areaName: string) =
 		enable_scroll_wheel_speed_control: (__oldValue, newValue) => {
 			sendExtensionOnlyMessage("scrollWheelSpeedControlChange", {
 				scrollWheelSpeedControlEnabled: newValue
-			});
-		},
-		enable_automatically_disable_closed_captions: (__oldValue, newValue) => {
-			sendExtensionOnlyMessage("automaticallyDisableClosedCaptionsChange", {
-				automaticallyDisableClosedCaptionsEnabled: newValue
 			});
 		},
 		enable_scroll_wheel_volume_control: (__oldValue, newValue) => {
