@@ -1,6 +1,5 @@
-/* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
 import { deepDarkPresets } from "@/src/deepDarkPresets";
-import { type FeatureFuncRecord, featureButtonFunctions } from "@/src/features";
+import { featureButtonFunctions, type FeatureFuncRecord } from "@/src/features";
 import {
 	disableAutomaticallyDisableClosedCaptions,
 	enableAutomaticallyDisableClosedCaptions
@@ -69,20 +68,20 @@ import volumeBoost, {
 	removeVolumeBoostButton
 } from "@/src/features/volumeBoost";
 import { i18nService } from "@/src/i18n";
-import { type ToggleFeatures, type ToggleIcon, getFeatureIcon, toggleFeatures } from "@/src/icons";
+import { getFeatureIcon, type ToggleFeatures, toggleFeatures, type ToggleIcon } from "@/src/icons";
 import {
 	type AllButtonNames,
 	type ButtonPlacement,
 	type ExtensionSendOnlyMessageMappings,
 	type FeatureToMultiButtonMap,
+	featureToMultiButtonsMap,
 	type KeysOfUnion,
 	type Messages,
 	type MultiButtonFeatureNames,
 	type MultiButtonNames,
 	type SingleButtonFeatureNames,
 	type SingleButtonNames,
-	type YouTubePlayerDiv,
-	featureToMultiButtonsMap
+	type YouTubePlayerDiv
 } from "@/src/types";
 import eventManager from "@/utils/EventManager";
 import {
@@ -200,33 +199,33 @@ const enableFeatures = () => {
 			const buttonName = featureToMultiButtonsMap.get(multiButtonFeatureName)?.at(-1);
 			if (!buttonName) continue;
 			switch (multiButtonFeatureName) {
-				case "playbackSpeedButtons": {
-					switch (button_placements[buttonName]) {
-						case "below_player":
-						case "player_controls_left":
-						case "feature_menu": {
-							await addDecreasePlaybackSpeedButton().then(addIncreasePlaybackSpeedButton);
-							break;
-						}
-						// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
-						case "player_controls_right": {
-							await addIncreasePlaybackSpeedButton().then(addDecreasePlaybackSpeedButton);
-							break;
-						}
-					}
-					break;
-				}
 				case "forwardRewindButtons": {
 					switch (button_placements[buttonName]) {
 						case "below_player":
-						case "player_controls_left":
-						case "feature_menu": {
+						case "feature_menu":
+						case "player_controls_left": {
 							await addRewindButton().then(addForwardButton);
 							break;
 						}
 						// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
 						case "player_controls_right": {
 							await addForwardButton().then(addRewindButton);
+							break;
+						}
+					}
+					break;
+				}
+				case "playbackSpeedButtons": {
+					switch (button_placements[buttonName]) {
+						case "below_player":
+						case "feature_menu":
+						case "player_controls_left": {
+							await addDecreasePlaybackSpeedButton().then(addIncreasePlaybackSpeedButton);
+							break;
+						}
+						// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
+						case "player_controls_right": {
+							await addIncreasePlaybackSpeedButton().then(addDecreasePlaybackSpeedButton);
 							break;
 						}
 					}
@@ -300,117 +299,194 @@ window.addEventListener("DOMContentLoaded", function () {
 				}
 				if (!message) return;
 				switch (message.type) {
-					case "hideOfficialArtistVideosFromHomePageChange": {
+					case "automaticallyDisableClosedCaptionsChange": {
 						const {
-							data: { hideOfficialArtistVideosFromHomePageEnabled }
+							data: { automaticallyDisableClosedCaptionsEnabled }
 						} = message;
-						if (hideOfficialArtistVideosFromHomePageEnabled) {
-							await enableHideOfficialArtistVideosFromHomePage();
+						if (automaticallyDisableClosedCaptionsEnabled) {
+							await enableAutomaticallyDisableClosedCaptions();
 						} else {
-							disableHideOfficialArtistVideosFromHomePage();
+							await disableAutomaticallyDisableClosedCaptions();
 						}
 						break;
 					}
-					case "playlistLengthChange": {
-						const {
-							data: { playlistLengthEnabled }
-						} = message;
-						if (playlistLengthEnabled) {
-							await enablePlaylistLength();
-						} else {
-							disablePlaylistLength();
+					case "automaticTheaterModeChange": {
+						// Get the player element
+						const playerContainer =
+							isWatchPage() ?
+								document.querySelector<HTMLDivElement>(
+									isNewYouTubeVideoLayout() ? "div#player-container.ytd-watch-grid" : "div#player-container.ytd-watch-flexy"
+								)
+							: isShortsPage() ? document.querySelector("div#shorts-player")
+							: null;
+						// If player element is not available, return
+						if (!playerContainer) return;
+						// Get the size button
+						const sizeButton = document.querySelector<HTMLButtonElement>("button.ytp-size-button");
+						// If the size button is not available return
+						if (!sizeButton) return;
+						sizeButton.click();
+						break;
+					}
+					case "buttonPlacementChange": {
+						const { data } = message;
+						const { multiButtonChanges, singleButtonChanges } = groupButtonChanges(data);
+						for (const [featureName, changes] of Object.entries(multiButtonChanges)) {
+							switch (featureName) {
+								case "forwardRewindButtons": {
+									for (const [buttonName, { new: newPlacement, old: oldPlacement }] of Object.entries(changes)) {
+										if (oldPlacement === newPlacement) continue;
+										const rewindButtonFuncs = getFeatureFunctions("rewindButton", oldPlacement);
+										const forwardButtonFuncs = getFeatureFunctions("forwardButton", oldPlacement);
+										switch (buttonName) {
+											case "forwardButton":
+											case "rewindButton": {
+												await forwardButtonFuncs.remove();
+												await rewindButtonFuncs.remove();
+												switch (newPlacement) {
+													case "below_player":
+													case "feature_menu":
+													case "player_controls_left": {
+														await rewindButtonFuncs.add().then(forwardButtonFuncs.add);
+														break;
+													}
+													// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
+													case "player_controls_right": {
+														await forwardButtonFuncs.add().then(rewindButtonFuncs.add);
+														break;
+													}
+												}
+											}
+										}
+									}
+									break;
+								}
+								case "playbackSpeedButtons": {
+									for (const [buttonName, { new: newPlacement, old: oldPlacement }] of Object.entries(changes)) {
+										if (oldPlacement === newPlacement) continue;
+										const increasePlaybackSpeedButtonFuncs = getFeatureFunctions("increasePlaybackSpeedButton", oldPlacement);
+										const decreasePlaybackSpeedButtonFuncs = getFeatureFunctions("decreasePlaybackSpeedButton", oldPlacement);
+										await decreasePlaybackSpeedButtonFuncs.remove();
+										await increasePlaybackSpeedButtonFuncs.remove();
+										switch (buttonName) {
+											case "decreasePlaybackSpeedButton":
+											case "increasePlaybackSpeedButton": {
+												switch (newPlacement) {
+													case "below_player":
+													case "feature_menu":
+													case "player_controls_left": {
+														await decreasePlaybackSpeedButtonFuncs.add().then(increasePlaybackSpeedButtonFuncs.add);
+														break;
+													}
+													// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
+													case "player_controls_right": {
+														await increasePlaybackSpeedButtonFuncs.add().then(decreasePlaybackSpeedButtonFuncs.add);
+														break;
+													}
+												}
+											}
+										}
+									}
+								}
+							}
+						}
+						for (const [featureName, { new: newPlacement, old: oldPlacement }] of Object.entries(singleButtonChanges)) {
+							if (oldPlacement === newPlacement) continue;
+							const featureFuncs = getFeatureFunctions(featureName, oldPlacement);
+							await featureFuncs.remove();
+							await featureFuncs.add();
 						}
 						break;
 					}
-					case "playlistLengthGetMethodChange":
-					case "playlistWatchTimeGetMethodChange": {
-						disablePlaylistLength();
-						await enablePlaylistLength();
+					case "copyTimestampUrlButtonChange": {
+						const {
+							data: { copyTimestampUrlButtonEnabled }
+						} = message;
+						if (copyTimestampUrlButtonEnabled) {
+							await addCopyTimestampUrlButton();
+						} else {
+							await removeCopyTimestampUrlButton();
+						}
 						break;
 					}
-					case "volumeBoostChange": {
+					case "customCSSChange": {
 						const {
-							data: { volumeBoostEnabled, volumeBoostMode }
+							data: { customCSSCode, customCSSEnabled }
 						} = message;
-						if (volumeBoostEnabled) {
-							if (volumeBoostMode === "global") {
-								await removeVolumeBoostButton();
-								await enableVolumeBoost();
+						if (customCSSEnabled) {
+							if (customCSSExists()) {
+								updateCustomCSS({ custom_css_code: customCSSCode });
 							} else {
-								disableVolumeBoost();
-								await addVolumeBoostButton();
+								await enableCustomCSS();
 							}
 						} else {
-							disableVolumeBoost();
-							if (volumeBoostMode === "per_video") {
-								await removeVolumeBoostButton();
-							}
+							disableCustomCSS();
 						}
 						break;
 					}
-					case "volumeBoostAmountChange": {
+					case "deepDarkThemeChange": {
 						const {
-							data: { volumeBoostAmount, volumeBoostEnabled, volumeBoostMode }
+							data: { deepDarkCustomThemeColors, deepDarkPreset, deepDarkThemeEnabled }
 						} = message;
-						switch (volumeBoostMode) {
-							case "global": {
-								if (!volumeBoostEnabled) return;
-								applyVolumeBoost(volumeBoostAmount);
-								break;
+						if (deepDarkThemeEnabled) {
+							if (deepDarkCSSExists()) {
+								updateDeepDarkCSS(
+									deepDarkPreset === "Custom" ? getDeepDarkCustomThemeStyle(deepDarkCustomThemeColors) : deepDarkPresets[deepDarkPreset]
+								);
+							} else {
+								await enableDeepDarkCSS();
 							}
-							case "per_video": {
-								const volumeBoostButton = getFeatureMenuItem("volumeBoostButton") ?? getFeatureButton("volumeBoostButton");
-								if (!volumeBoostButton) return;
-								const volumeBoostForVideoEnabled = volumeBoostButton.ariaChecked === "true";
-								if (volumeBoostForVideoEnabled) applyVolumeBoost(volumeBoostAmount);
-							}
+						} else {
+							disableDeepDarkCSS();
 						}
 						break;
 					}
-					case "playerSpeedChange": {
+					case "featureMenuOpenTypeChange": {
 						const {
-							data: { enableForcedPlaybackSpeed, playerSpeed }
+							data: { featureMenuOpenType }
+						} = message;
+						setupFeatureMenuEventListeners(featureMenuOpenType);
+						break;
+					}
+					case "forwardRewindButtonsChange": {
+						const {
+							data: { forwardRewindButtonsEnabled }
 						} = message;
 						const {
 							data: {
-								options: { playback_buttons_speed: playbackSpeedPerClick }
+								options: {
+									button_placements: { forwardButton: forwardButtonPlacement }
+								}
 							}
 						} = await waitForSpecificMessage("options", "request_data", "content");
-						if (enableForcedPlaybackSpeed && playerSpeed) {
-							await updatePlaybackSpeedButtonTooltip(
-								"increasePlaybackSpeedButton",
-								calculatePlaybackButtonSpeed(playerSpeed, playbackSpeedPerClick, "increase")
-							);
-							await updatePlaybackSpeedButtonTooltip(
-								"decreasePlaybackSpeedButton",
-								calculatePlaybackButtonSpeed(playerSpeed, playbackSpeedPerClick, "decrease")
-							);
-							await setPlayerSpeed(Number(playerSpeed));
-						} else if (!enableForcedPlaybackSpeed) {
-							restorePlayerSpeed();
-							const videoElement = document.querySelector<HTMLVideoElement>("video");
-							if (!videoElement) return;
-							const { playbackRate: currentSpeed } = videoElement;
-							await updatePlaybackSpeedButtonTooltip(
-								"increasePlaybackSpeedButton",
-								calculatePlaybackButtonSpeed(currentSpeed, playbackSpeedPerClick, "increase")
-							);
-							await updatePlaybackSpeedButtonTooltip(
-								"decreasePlaybackSpeedButton",
-								calculatePlaybackButtonSpeed(currentSpeed, playbackSpeedPerClick, "decrease")
-							);
+						await removeForwardButton();
+						await removeRewindButton();
+						if (forwardRewindButtonsEnabled) {
+							switch (forwardButtonPlacement) {
+								case "below_player":
+								case "feature_menu":
+								case "player_controls_left": {
+									await addRewindButton().then(addForwardButton);
+									break;
+								}
+								// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
+								case "player_controls_right": {
+									await addForwardButton().then(addRewindButton);
+									break;
+								}
+							}
+						} else {
+							await removeRewindButton();
+							await removeForwardButton();
 						}
 						break;
 					}
-					case "screenshotButtonChange": {
+					case "hideEndScreenCardsButtonChange": {
 						const {
-							data: { screenshotButtonEnabled }
+							data: { hideEndScreenCardsButtonEnabled }
 						} = message;
-						if (screenshotButtonEnabled) {
-							await addScreenshotButton();
-						} else {
-							await removeScreenshotButton();
-						}
+						if (hideEndScreenCardsButtonEnabled) await addHideEndScreenCardsButton();
+						else await removeHideEndScreenCardsButton();
 						break;
 					}
 					case "hideEndScreenCardsChange": {
@@ -445,177 +521,37 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
-					case "maximizeButtonChange": {
+					case "hideLiveStreamChatChange": {
 						const {
-							data: { maximizePlayerButtonEnabled }
+							data: { hideLiveStreamChatEnabled }
 						} = message;
-						if (maximizePlayerButtonEnabled) {
-							await addMaximizePlayerButton();
+						if (hideLiveStreamChatEnabled) {
+							await enableHideLiveStreamChat();
 						} else {
-							await removeMaximizePlayerButton();
-							const maximizePlayerButton = document.querySelector<HTMLButtonElement>("video.html5-main-video");
-							if (!maximizePlayerButton) return;
-							// Get the video element
-							const videoElement = document.querySelector<HTMLVideoElement>("video.html5-main-video");
-							// If video element is not available, return
-							if (!videoElement) return;
-							const videoContainer = document.querySelector<YouTubePlayerDiv>("video.html5-main-video");
-							if (!videoContainer) return;
-							if (videoContainer.classList.contains("maximized_video_container") && videoElement.classList.contains("maximized_video")) {
-								maximizePlayer();
-							}
+							await disableHideLiveStreamChat();
 						}
 						break;
 					}
-					case "videoHistoryChange": {
+					case "hideOfficialArtistVideosFromHomePageChange": {
 						const {
-							data: { videoHistoryEnabled }
+							data: { hideOfficialArtistVideosFromHomePageEnabled }
 						} = message;
-						if (videoHistoryEnabled) {
-							await setupVideoHistory();
+						if (hideOfficialArtistVideosFromHomePageEnabled) {
+							await enableHideOfficialArtistVideosFromHomePage();
 						} else {
-							eventManager.removeEventListeners("videoHistory");
+							disableHideOfficialArtistVideosFromHomePage();
 						}
 						break;
 					}
-					case "remainingTimeChange": {
+					case "hidePaidPromotionBannerChange": {
 						const {
-							data: { remainingTimeEnabled }
+							data: { hidePaidPromotionBannerEnabled }
 						} = message;
-						if (remainingTimeEnabled) {
-							await enableRemainingTime();
+						if (hidePaidPromotionBannerEnabled) {
+							await enableHidePaidPromotionBanner();
 						} else {
-							removeRemainingTimeDisplay();
+							disableHidePaidPromotionBanner();
 						}
-						break;
-					}
-					case "loopButtonChange": {
-						const {
-							data: { loopButtonEnabled }
-						} = message;
-						if (loopButtonEnabled) {
-							await addLoopButton();
-						} else {
-							await removeLoopButton();
-						}
-						break;
-					}
-					case "copyTimestampUrlButtonChange": {
-						const {
-							data: { copyTimestampUrlButtonEnabled }
-						} = message;
-						if (copyTimestampUrlButtonEnabled) {
-							await addCopyTimestampUrlButton();
-						} else {
-							await removeCopyTimestampUrlButton();
-						}
-						break;
-					}
-					case "forwardRewindButtonsChange": {
-						const {
-							data: { forwardRewindButtonsEnabled }
-						} = message;
-						const {
-							data: {
-								options: {
-									button_placements: { forwardButton: forwardButtonPlacement }
-								}
-							}
-						} = await waitForSpecificMessage("options", "request_data", "content");
-						await removeForwardButton();
-						await removeRewindButton();
-						if (forwardRewindButtonsEnabled) {
-							switch (forwardButtonPlacement) {
-								case "below_player":
-								case "player_controls_left":
-								case "feature_menu": {
-									await addRewindButton().then(addForwardButton);
-									break;
-								}
-								// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
-								case "player_controls_right": {
-									await addForwardButton().then(addRewindButton);
-									break;
-								}
-							}
-						} else {
-							await removeRewindButton();
-							await removeForwardButton();
-						}
-						break;
-					}
-					case "playbackSpeedButtonsChange": {
-						const {
-							data: { playbackSpeedButtonsEnabled }
-						} = message;
-						const {
-							data: {
-								options: {
-									button_placements: { decreasePlaybackSpeedButton: decreasePlaybackSpeedButtonPlacement }
-								}
-							}
-						} = await waitForSpecificMessage("options", "request_data", "content");
-						if (playbackSpeedButtonsEnabled) {
-							await removeDecreasePlaybackSpeedButton();
-							await removeIncreasePlaybackSpeedButton();
-							switch (decreasePlaybackSpeedButtonPlacement) {
-								case "below_player":
-								case "player_controls_left":
-								case "feature_menu": {
-									await addDecreasePlaybackSpeedButton().then(addIncreasePlaybackSpeedButton);
-									break;
-								}
-								// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
-								case "player_controls_right": {
-									await addIncreasePlaybackSpeedButton().then(addDecreasePlaybackSpeedButton);
-									break;
-								}
-							}
-						} else {
-							await removeDecreasePlaybackSpeedButton();
-							await removeIncreasePlaybackSpeedButton();
-						}
-						break;
-					}
-					case "scrollWheelVolumeControlChange": {
-						const {
-							data: { scrollWheelVolumeControlEnabled }
-						} = message;
-						if (scrollWheelVolumeControlEnabled) {
-							await adjustVolumeOnScrollWheel();
-						} else {
-							eventManager.removeEventListeners("scrollWheelVolumeControl");
-						}
-						break;
-					}
-					case "scrollWheelSpeedControlChange": {
-						const {
-							data: { scrollWheelSpeedControlEnabled }
-						} = message;
-						if (scrollWheelSpeedControlEnabled) {
-							await adjustSpeedOnScrollWheel();
-						} else {
-							eventManager.removeEventListeners("scrollWheelSpeedControl");
-						}
-						break;
-					}
-					case "rememberVolumeChange": {
-						const {
-							data: { rememberVolumeEnabled }
-						} = message;
-						if (rememberVolumeEnabled) {
-							await enableRememberVolume();
-						} else {
-							eventManager.removeEventListeners("rememberVolume");
-						}
-						break;
-					}
-					case "hideTranslateCommentChange": {
-						const {
-							data: { hideTranslateCommentEnabled }
-						} = message;
-						if (hideTranslateCommentEnabled) await enableHideTranslateComment();
-						else await disableHideTranslateComment();
 						break;
 					}
 					case "hideScrollBarChange": {
@@ -645,34 +581,12 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
-					case "hideEndScreenCardsButtonChange": {
+					case "hideTranslateCommentChange": {
 						const {
-							data: { hideEndScreenCardsButtonEnabled }
+							data: { hideTranslateCommentEnabled }
 						} = message;
-						if (hideEndScreenCardsButtonEnabled) await addHideEndScreenCardsButton();
-						else await removeHideEndScreenCardsButton();
-						break;
-					}
-					case "automaticallyDisableClosedCaptionsChange": {
-						const {
-							data: { automaticallyDisableClosedCaptionsEnabled }
-						} = message;
-						if (automaticallyDisableClosedCaptionsEnabled) {
-							await enableAutomaticallyDisableClosedCaptions();
-						} else {
-							await disableAutomaticallyDisableClosedCaptions();
-						}
-						break;
-					}
-					case "hideLiveStreamChatChange": {
-						const {
-							data: { hideLiveStreamChatEnabled }
-						} = message;
-						if (hideLiveStreamChatEnabled) {
-							await enableHideLiveStreamChat();
-						} else {
-							await disableHideLiveStreamChat();
-						}
+						if (hideTranslateCommentEnabled) await enableHideTranslateComment();
+						else await disableHideTranslateComment();
 						break;
 					}
 					case "languageChange": {
@@ -753,29 +667,37 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
-					case "automaticTheaterModeChange": {
-						// Get the player element
-						const playerContainer =
-							isWatchPage() ?
-								document.querySelector<HTMLDivElement>(
-									isNewYouTubeVideoLayout() ? "div#player-container.ytd-watch-grid" : "div#player-container.ytd-watch-flexy"
-								)
-							: isShortsPage() ? document.querySelector("div#shorts-player")
-							: null;
-						// If player element is not available, return
-						if (!playerContainer) return;
-						// Get the size button
-						const sizeButton = document.querySelector<HTMLButtonElement>("button.ytp-size-button");
-						// If the size button is not available return
-						if (!sizeButton) return;
-						sizeButton.click();
+					case "loopButtonChange": {
+						const {
+							data: { loopButtonEnabled }
+						} = message;
+						if (loopButtonEnabled) {
+							await addLoopButton();
+						} else {
+							await removeLoopButton();
+						}
 						break;
 					}
-					case "featureMenuOpenTypeChange": {
+					case "maximizeButtonChange": {
 						const {
-							data: { featureMenuOpenType }
+							data: { maximizePlayerButtonEnabled }
 						} = message;
-						setupFeatureMenuEventListeners(featureMenuOpenType);
+						if (maximizePlayerButtonEnabled) {
+							await addMaximizePlayerButton();
+						} else {
+							await removeMaximizePlayerButton();
+							const maximizePlayerButton = document.querySelector<HTMLButtonElement>("video.html5-main-video");
+							if (!maximizePlayerButton) return;
+							// Get the video element
+							const videoElement = document.querySelector<HTMLVideoElement>("video.html5-main-video");
+							// If video element is not available, return
+							if (!videoElement) return;
+							const videoContainer = document.querySelector<YouTubePlayerDiv>("video.html5-main-video");
+							if (!videoContainer) return;
+							if (videoContainer.classList.contains("maximized_video_container") && videoElement.classList.contains("maximized_video")) {
+								maximizePlayer();
+							}
+						}
 						break;
 					}
 					case "openTranscriptButtonChange": {
@@ -800,6 +722,124 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
+					case "pauseBackgroundPlayersChange": {
+						const {
+							data: { pauseBackgroundPlayersEnabled }
+						} = message;
+						if (pauseBackgroundPlayersEnabled) {
+							await enablePauseBackgroundPlayers();
+						} else {
+							disablePauseBackgroundPlayers();
+						}
+						break;
+					}
+					case "playbackSpeedButtonsChange": {
+						const {
+							data: { playbackSpeedButtonsEnabled }
+						} = message;
+						const {
+							data: {
+								options: {
+									button_placements: { decreasePlaybackSpeedButton: decreasePlaybackSpeedButtonPlacement }
+								}
+							}
+						} = await waitForSpecificMessage("options", "request_data", "content");
+						if (playbackSpeedButtonsEnabled) {
+							await removeDecreasePlaybackSpeedButton();
+							await removeIncreasePlaybackSpeedButton();
+							switch (decreasePlaybackSpeedButtonPlacement) {
+								case "below_player":
+								case "feature_menu":
+								case "player_controls_left": {
+									await addDecreasePlaybackSpeedButton().then(addIncreasePlaybackSpeedButton);
+									break;
+								}
+								// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
+								case "player_controls_right": {
+									await addIncreasePlaybackSpeedButton().then(addDecreasePlaybackSpeedButton);
+									break;
+								}
+							}
+						} else {
+							await removeDecreasePlaybackSpeedButton();
+							await removeIncreasePlaybackSpeedButton();
+						}
+						break;
+					}
+					case "playerSpeedChange": {
+						const {
+							data: { enableForcedPlaybackSpeed, playerSpeed }
+						} = message;
+						const {
+							data: {
+								options: { playback_buttons_speed: playbackSpeedPerClick }
+							}
+						} = await waitForSpecificMessage("options", "request_data", "content");
+						if (enableForcedPlaybackSpeed && playerSpeed) {
+							await updatePlaybackSpeedButtonTooltip(
+								"increasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(playerSpeed, playbackSpeedPerClick, "increase")
+							);
+							await updatePlaybackSpeedButtonTooltip(
+								"decreasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(playerSpeed, playbackSpeedPerClick, "decrease")
+							);
+							await setPlayerSpeed(Number(playerSpeed));
+						} else if (!enableForcedPlaybackSpeed) {
+							restorePlayerSpeed();
+							const videoElement = document.querySelector<HTMLVideoElement>("video");
+							if (!videoElement) return;
+							const { playbackRate: currentSpeed } = videoElement;
+							await updatePlaybackSpeedButtonTooltip(
+								"increasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(currentSpeed, playbackSpeedPerClick, "increase")
+							);
+							await updatePlaybackSpeedButtonTooltip(
+								"decreasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(currentSpeed, playbackSpeedPerClick, "decrease")
+							);
+						}
+						break;
+					}
+					case "playlistLengthChange": {
+						const {
+							data: { playlistLengthEnabled }
+						} = message;
+						if (playlistLengthEnabled) {
+							await enablePlaylistLength();
+						} else {
+							disablePlaylistLength();
+						}
+						break;
+					}
+					case "playlistLengthGetMethodChange":
+					case "playlistWatchTimeGetMethodChange": {
+						disablePlaylistLength();
+						await enablePlaylistLength();
+						break;
+					}
+					case "remainingTimeChange": {
+						const {
+							data: { remainingTimeEnabled }
+						} = message;
+						if (remainingTimeEnabled) {
+							await enableRemainingTime();
+						} else {
+							removeRemainingTimeDisplay();
+						}
+						break;
+					}
+					case "rememberVolumeChange": {
+						const {
+							data: { rememberVolumeEnabled }
+						} = message;
+						if (rememberVolumeEnabled) {
+							await enableRememberVolume();
+						} else {
+							eventManager.removeEventListeners("rememberVolume");
+						}
+						break;
+					}
 					case "removeRedirectChange": {
 						const {
 							data: { removeRedirectEnabled }
@@ -809,14 +849,36 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
-					case "pauseBackgroundPlayersChange": {
+					case "screenshotButtonChange": {
 						const {
-							data: { pauseBackgroundPlayersEnabled }
+							data: { screenshotButtonEnabled }
 						} = message;
-						if (pauseBackgroundPlayersEnabled) {
-							await enablePauseBackgroundPlayers();
+						if (screenshotButtonEnabled) {
+							await addScreenshotButton();
 						} else {
-							disablePauseBackgroundPlayers();
+							await removeScreenshotButton();
+						}
+						break;
+					}
+					case "scrollWheelSpeedControlChange": {
+						const {
+							data: { scrollWheelSpeedControlEnabled }
+						} = message;
+						if (scrollWheelSpeedControlEnabled) {
+							await adjustSpeedOnScrollWheel();
+						} else {
+							eventManager.removeEventListeners("scrollWheelSpeedControl");
+						}
+						break;
+					}
+					case "scrollWheelVolumeControlChange": {
+						const {
+							data: { scrollWheelVolumeControlEnabled }
+						} = message;
+						if (scrollWheelVolumeControlEnabled) {
+							await adjustVolumeOnScrollWheel();
+						} else {
+							eventManager.removeEventListeners("scrollWheelVolumeControl");
 						}
 						break;
 					}
@@ -831,6 +893,17 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
+					case "shortsAutoScrollChange": {
+						const {
+							data: { shortsAutoScrollEnabled }
+						} = message;
+						if (shortsAutoScrollEnabled) {
+							await enableShortsAutoScroll();
+						} else {
+							disableShortsAutoScroll();
+						}
+						break;
+					}
 					case "skipContinueWatchingChange": {
 						const {
 							data: { skipContinueWatchingEnabled }
@@ -840,127 +913,53 @@ window.addEventListener("DOMContentLoaded", function () {
 						}
 						break;
 					}
-					case "hidePaidPromotionBannerChange": {
+					case "videoHistoryChange": {
 						const {
-							data: { hidePaidPromotionBannerEnabled }
+							data: { videoHistoryEnabled }
 						} = message;
-						if (hidePaidPromotionBannerEnabled) {
-							await enableHidePaidPromotionBanner();
+						if (videoHistoryEnabled) {
+							await setupVideoHistory();
 						} else {
-							disableHidePaidPromotionBanner();
+							eventManager.removeEventListeners("videoHistory");
 						}
 						break;
 					}
-					case "deepDarkThemeChange": {
+					case "volumeBoostAmountChange": {
 						const {
-							data: { deepDarkCustomThemeColors, deepDarkPreset, deepDarkThemeEnabled }
+							data: { volumeBoostAmount, volumeBoostEnabled, volumeBoostMode }
 						} = message;
-						if (deepDarkThemeEnabled) {
-							if (deepDarkCSSExists()) {
-								updateDeepDarkCSS(
-									deepDarkPreset === "Custom" ? getDeepDarkCustomThemeStyle(deepDarkCustomThemeColors) : deepDarkPresets[deepDarkPreset]
-								);
+						switch (volumeBoostMode) {
+							case "global": {
+								if (!volumeBoostEnabled) return;
+								applyVolumeBoost(volumeBoostAmount);
+								break;
+							}
+							case "per_video": {
+								const volumeBoostButton = getFeatureMenuItem("volumeBoostButton") ?? getFeatureButton("volumeBoostButton");
+								if (!volumeBoostButton) return;
+								const volumeBoostForVideoEnabled = volumeBoostButton.ariaChecked === "true";
+								if (volumeBoostForVideoEnabled) applyVolumeBoost(volumeBoostAmount);
+							}
+						}
+						break;
+					}
+					case "volumeBoostChange": {
+						const {
+							data: { volumeBoostEnabled, volumeBoostMode }
+						} = message;
+						if (volumeBoostEnabled) {
+							if (volumeBoostMode === "global") {
+								await removeVolumeBoostButton();
+								await enableVolumeBoost();
 							} else {
-								await enableDeepDarkCSS();
+								disableVolumeBoost();
+								await addVolumeBoostButton();
 							}
 						} else {
-							disableDeepDarkCSS();
-						}
-						break;
-					}
-					case "customCSSChange": {
-						const {
-							data: { customCSSCode, customCSSEnabled }
-						} = message;
-						if (customCSSEnabled) {
-							if (customCSSExists()) {
-								updateCustomCSS({ custom_css_code: customCSSCode });
-							} else {
-								await enableCustomCSS();
+							disableVolumeBoost();
+							if (volumeBoostMode === "per_video") {
+								await removeVolumeBoostButton();
 							}
-						} else {
-							disableCustomCSS();
-						}
-						break;
-					}
-					case "buttonPlacementChange": {
-						const { data } = message;
-						const { multiButtonChanges, singleButtonChanges } = groupButtonChanges(data);
-						for (const [featureName, changes] of Object.entries(multiButtonChanges)) {
-							switch (featureName) {
-								case "playbackSpeedButtons": {
-									for (const [buttonName, { new: newPlacement, old: oldPlacement }] of Object.entries(changes)) {
-										if (oldPlacement === newPlacement) continue;
-										const increasePlaybackSpeedButtonFuncs = getFeatureFunctions("increasePlaybackSpeedButton", oldPlacement);
-										const decreasePlaybackSpeedButtonFuncs = getFeatureFunctions("decreasePlaybackSpeedButton", oldPlacement);
-										await decreasePlaybackSpeedButtonFuncs.remove();
-										await increasePlaybackSpeedButtonFuncs.remove();
-										switch (buttonName) {
-											case "increasePlaybackSpeedButton":
-											case "decreasePlaybackSpeedButton": {
-												switch (newPlacement) {
-													case "below_player":
-													case "player_controls_left":
-													case "feature_menu": {
-														await decreasePlaybackSpeedButtonFuncs.add().then(increasePlaybackSpeedButtonFuncs.add);
-														break;
-													}
-													// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
-													case "player_controls_right": {
-														await increasePlaybackSpeedButtonFuncs.add().then(decreasePlaybackSpeedButtonFuncs.add);
-														break;
-													}
-												}
-											}
-										}
-									}
-									break;
-								}
-								case "forwardRewindButtons": {
-									for (const [buttonName, { new: newPlacement, old: oldPlacement }] of Object.entries(changes)) {
-										if (oldPlacement === newPlacement) continue;
-										const rewindButtonFuncs = getFeatureFunctions("rewindButton", oldPlacement);
-										const forwardButtonFuncs = getFeatureFunctions("forwardButton", oldPlacement);
-										switch (buttonName) {
-											case "forwardButton":
-											case "rewindButton": {
-												await forwardButtonFuncs.remove();
-												await rewindButtonFuncs.remove();
-												switch (newPlacement) {
-													case "below_player":
-													case "player_controls_left":
-													case "feature_menu": {
-														await rewindButtonFuncs.add().then(forwardButtonFuncs.add);
-														break;
-													}
-													// Because of how the right controls are placed in the DOM, we need to add the buttons in reverse order
-													case "player_controls_right": {
-														await forwardButtonFuncs.add().then(rewindButtonFuncs.add);
-														break;
-													}
-												}
-											}
-										}
-									}
-								}
-							}
-						}
-						for (const [featureName, { new: newPlacement, old: oldPlacement }] of Object.entries(singleButtonChanges)) {
-							if (oldPlacement === newPlacement) continue;
-							const featureFuncs = getFeatureFunctions(featureName, oldPlacement);
-							await featureFuncs.remove();
-							await featureFuncs.add();
-						}
-						break;
-					}
-					case "shortsAutoScrollChange": {
-						const {
-							data: { shortsAutoScrollEnabled }
-						} = message;
-						if (shortsAutoScrollEnabled) {
-							await enableShortsAutoScroll();
-						} else {
-							disableShortsAutoScroll();
 						}
 						break;
 					}
@@ -982,7 +981,6 @@ window.onbeforeunload = function () {
 window.addEventListener("error", (event) => {
 	event.preventDefault();
 	const {
-		// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 		error: { stack: errorLine }
 	} = event;
 	browserColorLog(formatError(event.error) + "\nAt: " + errorLine, "FgRed");
